463 allocation page calendar frontend enhancement

+++ app/Spring-Microservices/applicationservice/src/main/java/com/infinity/applicationservice/dtos/Allocations/AllocationRequest.java
@@ -6,6 +6,6 @@ public record AllocationRequest(
         Long studentId,
         Long applicationId,
         TaskType task,
-        int hours,
+        double hours,
         Long sectionId) {
 }
